DOC-200. Moved not needed interface
