+++ datasafe-metainfo/datasafe-metainfo-version-impl/src/main/java/de/adorsys/datasafe/metainfo/version/impl/version/VersionEncoderDecoder.java
@@ -1,4 +1,4 @@
-package de.adorsys.datasafe.metainfo.version.api.version;
+package de.adorsys.datasafe.metainfo.version.impl.version;
 
 import de.adorsys.datasafe.types.api.resource.Uri;
 import de.adorsys.datasafe.types.api.resource.VersionedUri;
